VIB-42 Add cmake function to simplify adding tests
Added new cmake function which adds an executable for either
a unit or an integration test, if used. This provides a new
generic function and makes the old two functions make_gtest
and make_integration_test obsolete. Using the function one
has to specify the executable name, all SOURCES, all LIBS
which shall be linked, as well as all CONFIGS that are used
by the test and that shall be copied to the corresponding
build binary directory.