-- Minor bugfix with multiple comAdapter + logger
If more than one comAdapter with the same participantName gets
created, spdlog might throw an exception because a logger with
that name does already exist. Furthermore, set_default_logger
drops the last default logger from the spdlog registry which
leads to unexpected behaviour in the case of the comAdapter.

The solution is to drop the logger of the comAdapter immediately
after creating and not to use the default logger.