AFTMAGT264 add new IEthController::SendFrame(EthFrame) methods
We deprecate the SendMessage(EthMessage) in favor of the new SendFrame
methods.
The EthMessage supports a user-defined timestamp, which might lead to
confusion when using the new time provider API.
In the future, we should only rely on the SendFrame() methods with
explicit timestamps or implicitly derived timestamps from the time
provider.