AFTMAGT-312 Add backward compatibility
Ensure that we are compatible with VIB-2.0 registry messages.
The KnownParticipant contains a vector of VasioPeerInfo structs,
which makes it difficult to extend it.

We added a new VAsioPeerUri struct that is encoded additionally, outside
of the peer info struct. This allows legacy clients to parse the
existing peer infos with a known, fixed data size.

TODO in the future we should make the ser/des of vasio more resilient to
additions, changes etc.