SILKIT-944: Add partial implementation of the C++ API using the C API
This change introduces a partial implementation of the C++ API using the C API. The intent is to cover the C API using the same (integration) tests as the C++ API.

Some tests still require using the internal C++ interfaces (e.g. the RPC ITests require access to the ServiceDiscovery).
- Build various integration tests using the C-API wrapper
- Add Ethernet Demo variant using the C-API wrapper if building from source (not available for demo-builds from the release package)