SILKIT-1036: ASan: Fix heap-buffer-overflow (read) in ITest_ParticipantVersion
The issue was caused by an invalid assumption. The code in SerializedMessage which tried to peek the RegistryMsgHeader field of any registry/handshake message did not account for the fact, that the ParticipantAnnouncementReply in version 3.0 does not contain the RegistryMsgHeader.

The 3.0 protocol is only used by participants with versions below 3.99.23 (i.e. pre-initial-release of SIL Kit).