SILKIT-1036: TSan: Fix data-race on LifecycleService::_status
This fix introduces a workaround for the implementation of the ILifecycleService::Status() method which returns a 'const ParticipantStatus &'. It can now be called safely, but the returned reference must only be used until the next call to ILifecycleService::Status().

In particular, the ILifecycleService::Status() method must not be called concurrently.