config: move to rapidyaml and remove yaml-cpp 🚀 (#207)
Add rapidyaml 0.9.0 as a header-only dependency, and remove yaml-cpp.

I refactored the code into a YamlReader and a YamlWriter, and tried to remove as much of rapidyaml from the parsing interface.
The `schemaVersion` field in yaml files is deprecated in favor of `SchemaVersion`  with capital s. 
The version so far was not validated when loading a config file.

SILKIT-1739